requirements.txt
# Core AI and ML libraries
google-generativeai
faiss-cpu
numpy

# FastAPI and web framework
fastapi
uvicorn[standard]
python-multipart

# Environment and configuration
python-dotenv

# PDF Processing
PyPDF2
pdfplumber

# Image Processing and OCR
Pillow
pytesseract

# Word Document Processing
python-docx
docx2txt

# PowerPoint Processing
python-pptx

# Excel Processing
pandas
openpyxl
xlrd

# HTML/XML Processing
beautifulsoup4
lxml

# Text Processing
chardet

# Development and testing
pytest
pytest-asyncio
httpx

fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
PyPDF2
requests
aiofiles
requests
python-dotenv
google-generativeai
